All tests green on windows.

+++ src/main/java/com/orientechnologies/teleporter/util/OFileManager.java
@@ -30,14 +30,18 @@ public class OFileManager {
   public static void deleteResource(String resourcePath) throws IOException {
 
     File currentFile = new File(resourcePath);
-    if (currentFile.isDirectory()) {
-      File[] innerFiles = currentFile.listFiles();
-      for (File file : innerFiles) {
-        deleteResource(file.getCanonicalPath());
+    if(currentFile.exists()) {
+      if (currentFile.isDirectory()) {
+        File[] innerFiles = currentFile.listFiles();
+        for (File file : innerFiles) {
+          deleteResource(file.getCanonicalPath());
+        }
+        currentFile.delete();
+      } else {
+        if (!currentFile.delete())
+          throw new IOException();
       }
     }
-    if (!currentFile.delete())
-      throw new IOException();
   }
 
   public static void extractAll(String inputArchiveFilePath, String outputFolderPath) throws IOException {
